initial move of hypothesis/parametric tests into their own parallel dir
